+++ test/bigtest/network/index.js
@@ -1,4 +1,7 @@
-import { camelize } from '@bigtest/mirage';
+import {
+  camelize,
+  underscore,
+} from 'inflected';
 
 // auto-import all mirage submodules
 const req = require.context('./', true, /\.js$/);
@@ -6,7 +9,7 @@ const modules = req.keys().reduce((acc, modulePath) => {
   const [, moduleType, moduleName] = modulePath.split('/');
 
   if (moduleName) {
-    const moduleKey = camelize(moduleName.replace('.js', ''));
+    const moduleKey = camelize(underscore(moduleName.replace('.js', '')), false);
 
     return Object.assign(acc, {
       [moduleType]: {